requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
outcome==1.2.0
PySocks==1.7.1
python-dateutil==2.8.2
requests==2.28.1
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
wsproto==1.2.0
